feat: cc-0043 divide auth and user into separate modules
