+++ src/auth/auth.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from "@nestjs/testing";
+import { user } from "../user/user.testData";
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
-import { user } from "./auth.testData";
 
 describe("AuthController", () => {
   let controller: AuthController;
