+++ src/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from "@nestjs/common";
 import { AuthService } from "./auth.service";
 import { AuthController } from "./auth.controller";
 import { MongooseModule } from "@nestjs/mongoose";
-import { User, UserSchema } from "./schemas/user.schema";
+import { User, UserSchema } from "../user/schemas/user.schema";
 
 @Module({
   imports: [
