+++ src/user/user.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from "@nestjs/common";
+
+@Controller("user")
+export class UserController {}
