+++ src/user/user.module.ts
@@ -1,0 +1,19 @@
+import { Module } from "@nestjs/common";
+import { MongooseModule } from "@nestjs/mongoose";
+import { User, UserSchema } from "./schemas/user.schema";
+import { UserController } from "./user.controller";
+import { UserService } from "./user.service";
+
+@Module({
+  imports: [
+    MongooseModule.forFeature([
+      {
+        name: User.name,
+        schema: UserSchema,
+      },
+    ]),
+  ],
+  providers: [UserService],
+  controllers: [UserController],
+})
+export class UserModule {}
